+++ yash-syntax/src/parser/lex/arith.rs
@@ -21,6 +21,7 @@ use crate::parser::core::Result;
 use crate::parser::error::Error;
 use crate::parser::error::SyntaxError;
 use crate::source::Location;
+use crate::source::Span;
 use crate::syntax::Text;
 use crate::syntax::TextUnit;
 use std::future::Future;
@@ -36,10 +37,10 @@ impl Lexer<'_> {
     /// to the closing `))` (inclusive). Otherwise, no characters are consumed
     /// and the return value is `Ok(None)`.
     ///
-    /// The `location` parameter should be the location of the initial `$`. It
-    /// is used to construct the result, but this function does not check if it
-    /// actually is a location of `$`.
-    pub async fn arithmetic_expansion(&mut self, location: Location) -> Result<Option<TextUnit>> {
+    /// The `start` parameter should be the index for the initial `$`. It is
+    /// used to construct the result, but this function does not check if it
+    /// actually points to the `$`.
+    pub async fn arithmetic_expansion(&mut self, start: usize) -> Result<Option<TextUnit>> {
         let index = self.index();
 
         // Part 1: Parse `((`
@@ -64,7 +65,11 @@ impl Lexer<'_> {
             Some(sc) if sc == ')' => self.consume_char(),
             Some(_) => unreachable!(),
             None => {
-                let opening_location = location;
+                let Span { code, range } = self.span(start..start);
+                let opening_location = Location {
+                    code,
+                    index: range.start,
+                };
                 let cause = SyntaxError::UnclosedArith { opening_location }.into();
                 let location = self.location().await?.clone();
                 return Err(Error { cause, location });
@@ -77,14 +82,19 @@ impl Lexer<'_> {
                 return Ok(None);
             }
             None => {
-                let opening_location = location;
+                let Span { code, range } = self.span(start..start);
+                let opening_location = Location {
+                    code,
+                    index: range.start,
+                };
                 let cause = SyntaxError::UnclosedArith { opening_location }.into();
                 let location = self.location().await?.clone();
                 return Err(Error { cause, location });
             }
         }
 
-        Ok(Some(TextUnit::Arith { content, location }))
+        let span = self.span(start..self.index());
+        Ok(Some(TextUnit::Arith { content, span }))
     }
 }
 
@@ -100,18 +110,17 @@ mod tests {
 
     #[test]
     fn lexer_arithmetic_expansion_empty() {
-        let mut lexer = Lexer::from_memory("(());", Source::Unknown);
-        let location = Location::dummy("X");
+        let mut lexer = Lexer::from_memory("$(());", Source::Unknown);
+        block_on(lexer.peek_char()).unwrap();
+        lexer.consume_char();
 
-        let result = block_on(lexer.arithmetic_expansion(location))
-            .unwrap()
-            .unwrap();
-        assert_matches!(result, TextUnit::Arith { content, location } => {
+        let result = block_on(lexer.arithmetic_expansion(0)).unwrap().unwrap();
+        assert_matches!(result, TextUnit::Arith { content, span } => {
             assert_eq!(content.0, []);
-            assert_eq!(*location.code.value.borrow(), "X");
-            assert_eq!(location.code.start_line_number.get(), 1);
-            assert_eq!(location.code.source, Source::Unknown);
-            assert_eq!(location.index, 0);
+            assert_eq!(*span.code.value.borrow(), "$(());");
+            assert_eq!(span.code.start_line_number.get(), 1);
+            assert_eq!(span.code.source, Source::Unknown);
+            assert_eq!(span.range, 0..5);
         });
 
         assert_eq!(block_on(lexer.peek_char()), Ok(Some(';')));
@@ -120,25 +129,23 @@ mod tests {
     #[test]
     fn lexer_arithmetic_expansion_none() {
         let mut lexer = Lexer::from_memory("( foo bar )baz", Source::Unknown);
-        let location = Location::dummy("Y");
-        assert_eq!(block_on(lexer.arithmetic_expansion(location)), Ok(None));
+        assert_eq!(block_on(lexer.arithmetic_expansion(0)), Ok(None));
         assert_eq!(block_on(lexer.peek_char()), Ok(Some('(')));
     }
 
     #[test]
     fn lexer_arithmetic_expansion_line_continuations() {
-        let mut lexer = Lexer::from_memory("(\\\n\\\n(\\\n)\\\n\\\n);", Source::Unknown);
-        let location = Location::dummy("X");
+        let mut lexer = Lexer::from_memory("$(\\\n\\\n(\\\n)\\\n\\\n);", Source::Unknown);
+        block_on(lexer.peek_char()).unwrap();
+        lexer.consume_char();
 
-        let result = block_on(lexer.arithmetic_expansion(location))
-            .unwrap()
-            .unwrap();
-        assert_matches!(result, TextUnit::Arith { content, location } => {
+        let result = block_on(lexer.arithmetic_expansion(0)).unwrap().unwrap();
+        assert_matches!(result, TextUnit::Arith { content, span } => {
             assert_eq!(content.0, []);
-            assert_eq!(*location.code.value.borrow(), "X");
-            assert_eq!(location.code.start_line_number.get(), 1);
-            assert_eq!(location.code.source, Source::Unknown);
-            assert_eq!(location.index, 0);
+            assert_eq!(*span.code.value.borrow(), "$(\\\n\\\n(\\\n)\\\n\\\n);");
+            assert_eq!(span.code.start_line_number.get(), 1);
+            assert_eq!(span.code.source, Source::Unknown);
+            assert_eq!(span.range, 0..15);
         });
 
         assert_eq!(block_on(lexer.peek_char()), Ok(Some(';')));
@@ -146,13 +153,12 @@ mod tests {
 
     #[test]
     fn lexer_arithmetic_expansion_escapes() {
-        let mut lexer = Lexer::from_memory(r#"((\\\"\`\$));"#, Source::Unknown);
-        let location = Location::dummy("X");
+        let mut lexer = Lexer::from_memory(r#"$((\\\"\`\$));"#, Source::Unknown);
+        block_on(lexer.peek_char()).unwrap();
+        lexer.consume_char();
 
-        let result = block_on(lexer.arithmetic_expansion(location))
-            .unwrap()
-            .unwrap();
-        assert_matches!(result, TextUnit::Arith { content, location } => {
+        let result = block_on(lexer.arithmetic_expansion(0)).unwrap().unwrap();
+        assert_matches!(result, TextUnit::Arith { content, span } => {
             assert_eq!(
                 content.0,
                 [
@@ -163,10 +169,10 @@ mod tests {
                     Backslashed('$')
                 ]
             );
-            assert_eq!(*location.code.value.borrow(), "X");
-            assert_eq!(location.code.start_line_number.get(), 1);
-            assert_eq!(location.code.source, Source::Unknown);
-            assert_eq!(location.index, 0);
+            assert_eq!(*span.code.value.borrow(), r#"$((\\\"\`\$));"#);
+            assert_eq!(span.code.start_line_number.get(), 1);
+            assert_eq!(span.code.source, Source::Unknown);
+            assert_eq!(span.range, 0..13);
         });
 
         assert_eq!(block_on(lexer.peek_char()), Ok(Some(';')));
@@ -174,47 +180,48 @@ mod tests {
 
     #[test]
     fn lexer_arithmetic_expansion_unclosed_first() {
-        let mut lexer = Lexer::from_memory("((1", Source::Unknown);
-        let location = Location::dummy("Z");
+        let mut lexer = Lexer::from_memory("$((1", Source::Unknown);
+        block_on(lexer.peek_char()).unwrap();
+        lexer.consume_char();
 
-        let e = block_on(lexer.arithmetic_expansion(location)).unwrap_err();
+        let e = block_on(lexer.arithmetic_expansion(0)).unwrap_err();
         assert_matches!(e.cause,
             ErrorCause::Syntax(SyntaxError::UnclosedArith { opening_location }) => {
-            assert_eq!(*opening_location.code.value.borrow(), "Z");
+            assert_eq!(*opening_location.code.value.borrow(), "$((1");
             assert_eq!(opening_location.code.start_line_number.get(), 1);
             assert_eq!(opening_location.code.source, Source::Unknown);
             assert_eq!(opening_location.index, 0);
         });
-        assert_eq!(*e.location.code.value.borrow(), "((1");
-        assert_eq!(e.location.code.start_line_number.get(), 1);
-        assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.index, 3);
-    }
-
-    #[test]
-    fn lexer_arithmetic_expansion_unclosed_second() {
-        let mut lexer = Lexer::from_memory("((1)", Source::Unknown);
-        let location = Location::dummy("Z");
-
-        let e = block_on(lexer.arithmetic_expansion(location)).unwrap_err();
-        assert_matches!(e.cause,
-            ErrorCause::Syntax(SyntaxError::UnclosedArith { opening_location }) => {
-            assert_eq!(*opening_location.code.value.borrow(), "Z");
-            assert_eq!(opening_location.code.start_line_number.get(), 1);
-            assert_eq!(opening_location.code.source, Source::Unknown);
-            assert_eq!(opening_location.index, 0);
-        });
-        assert_eq!(*e.location.code.value.borrow(), "((1)");
+        assert_eq!(*e.location.code.value.borrow(), "$((1");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
         assert_eq!(e.location.index, 4);
     }
 
     #[test]
+    fn lexer_arithmetic_expansion_unclosed_second() {
+        let mut lexer = Lexer::from_memory("$((1)", Source::Unknown);
+        block_on(lexer.peek_char()).unwrap();
+        lexer.consume_char();
+
+        let e = block_on(lexer.arithmetic_expansion(0)).unwrap_err();
+        assert_matches!(e.cause,
+            ErrorCause::Syntax(SyntaxError::UnclosedArith { opening_location }) => {
+            assert_eq!(*opening_location.code.value.borrow(), "$((1)");
+            assert_eq!(opening_location.code.start_line_number.get(), 1);
+            assert_eq!(opening_location.code.source, Source::Unknown);
+            assert_eq!(opening_location.index, 0);
+        });
+        assert_eq!(*e.location.code.value.borrow(), "$((1)");
+        assert_eq!(e.location.code.start_line_number.get(), 1);
+        assert_eq!(e.location.code.source, Source::Unknown);
+        assert_eq!(e.location.index, 5);
+    }
+
+    #[test]
     fn lexer_arithmetic_expansion_unclosed_but_maybe_command_substitution() {
         let mut lexer = Lexer::from_memory("((1) ", Source::Unknown);
-        let location = Location::dummy("Z");
-        assert_eq!(block_on(lexer.arithmetic_expansion(location)), Ok(None));
+        assert_eq!(block_on(lexer.arithmetic_expansion(0)), Ok(None));
         assert_eq!(lexer.index(), 0);
     }
 }