+++ yash-syntax/CHANGELOG.md
@@ -16,8 +16,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - The following items' field `location: source::Location` to `span: source::Span`
+    - `syntax::TextUnit::Arith`
     - `syntax::TextUnit::CommandSubst`
     - `syntax::TextUnit::RawParam`
+- The following functions' parameter `location: source::Location` to `start:
+  usize`, with the return types modified accordingly
+    - `parser::Lexer::arithmetic_expansion`
+    - `parser::Lexer::raw_param`
 
 ## [0.3.0] - 2022-02-06